feat: move reports page out of team slug
